Updates and styles login page
