+++ src/pages/Login/LoginForm/index.styles.jsx
@@ -1,0 +1,4 @@
+import styled from "styled-components";
+import { BaseForm } from "../../../components/forms/BaseForm/index.styles";
+
+export const LoginForm = styled(BaseForm)``;
