+++ src/pages/Login/index.styles.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const LoginWrapper = styled.div`
+  margin: 0 auto;
+  max-width: ${({ theme }) => theme.screen["extra-small"]};
+
+  h1 {
+    text-align: center;
+    font-weight: normal;
+  }
+`;
